% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{Describe}
\alias{Describe}
\title{Summary Statistics}
\usage{
Describe(data, digits = 2)
}
\arguments{
\item{data}{vector of observations}

\item{digits}{integer deciding the number of decimal places}
}
\value{
A tibble of statistics, including min, max, mean, sd, Q25, Q50, Q75, kurt, Skew, n, na
}
\description{
Calculate useful statistics for an multivariate data.
}
\examples{
swan <- rGarch(len = 180)
Describe(tibble(a1 = swan, a2 = swan + 1), 2)
}
