% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos.R
\name{EuroCallOption}
\alias{EuroCallOption}
\title{The Value Function of European Call Option}
\usage{
EuroCallOption(x, K)
}
\arguments{
\item{x}{the stock price}

\item{K}{the strike price}
}
\value{
The value of European call option
}
\description{
With global variable K, the strike price, calculate the value of European call option.
}
\examples{
EuroCallOption(x = 2,K = 1)
}
