% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggacf.R
\name{ggacf}
\alias{ggacf}
\title{Plot the Acf Figure}
\usage{
ggacf(data, lag = 10)
}
\arguments{
\item{data}{vector of observations}

\item{lag}{the maximum lag to calculate the acf}
}
\value{
A ggplot figure of the acf
}
\description{
Plot the Acf figure with observations of a single variable beautifully.
}
\examples{
swan <- rGarch(len = 180)
ggacf(swan^2, 20)
}
