% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{gghistplot}
\alias{gghistplot}
\title{Plot the Histogram Figure}
\usage{
gghistplot(data, mapping = aes(x = x, y = stat(density)), bins = 10)
}
\arguments{
\item{data}{a tibble}

\item{mapping}{the mapping parameter of ggplot}

\item{bins}{the number of bins}
}
\value{
A histogram figure by ggplot
}
\description{
Plot the histgram figure beautifully with ggplot.
}
\examples{
date <- as.Date("2015-01-01") + days(0:180)
thero <- returns(rGbm("thero", date))[-1]
tthero <- tibble(x = date[-1], y = thero)
gghistplot(tthero, aes(x = thero, y = stat(density)), bins = 20)
}
