% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggacf.R
\name{ggpacf}
\alias{ggpacf}
\title{Plot the Pacf Figure}
\usage{
ggpacf(data, lag = 10)
}
\arguments{
\item{data}{vector of observations}

\item{lag}{the maximum lag to calculate the pacf}
}
\value{
A ggplot figure of the pacf
}
\description{
Plot the Pacf figure with observations of a single variable beautifully.
}
\examples{
swan <- rGarch(len = 180)
ggpacf(swan^2, 20)
}
