% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos.R
\name{NormChf}
\alias{NormChf}
\title{The Characteristic Function of Normal Distribution}
\usage{
NormChf(u, mu = 0, sigma = 1)
}
\arguments{
\item{u}{observation}

\item{mu}{the \code{mu} parameter}

\item{sigma}{the \code{sigma} parameter}
}
\value{
The value of Characteristic Function
}
\description{
The Characteristic Function of Normal Distribution
}
\examples{
NormChf(1)
}
