% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos.R
\name{PdfSinglePlot}
\alias{PdfSinglePlot}
\title{Plot the Probability Density Function}
\usage{
PdfSinglePlot(data, mapping = aes(x, y))
}
\arguments{
\item{data}{a tiible contains x and y}

\item{mapping}{the mapping parameter of ggplot}
}
\value{
A ggplot figure of the probability density function
}
\description{
Plot the p.d.f function for the univariate distribution with x and y.
}
\examples{
N <- 32
x <- seq(-5, 5, by = 10 / (32 - 1))
a <- -6.0
b <- 6.0
f_x <- CosPdfRecovery(x, StNormChf, N, a, b)
tnorm <- tibble(x = x, y = f_x)
PdfSinglePlot(tnorm)
}
