% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbm.R
\name{rGbm}
\alias{rGbm}
\title{Simulate prices series of stocks}
\usage{
rGbm(name, time, start = 100, mu = 0.01, sigma = 0.02)
}
\arguments{
\item{name}{vector of series names}

\item{time}{vector of time, must be a "Date" type variable}

\item{start}{vector of start positions}

\item{mu}{vector of \code{mu}}

\item{sigma}{vector of \code{sigma}}
}
\value{
a simulated multivariate GBM series
}
\description{
Simulate an multivariate series following Geometric Brownian Motion (GBM)
}
\examples{
date <- as.Date("2015-01-01") + days(0:29)
rGbm(c("bear", "tiger", "swan"), date)
}
