% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbm.R
\name{rGbmSingle}
\alias{rGbmSingle}
\title{Simulate a single stock price series}
\usage{
rGbmSingle(len, start = 100, mu = 0.01, sigma = 0.02)
}
\arguments{
\item{len}{the length}

\item{start}{the start position}

\item{mu}{the \code{mu} parameter of GBM}

\item{sigma}{the \code{sigma} parameter of GBM}
}
\value{
a simulated univariate GBM series
}
\description{
Simulate an univariate series following Geometric Brownian Motion (GBM).
}
\examples{
rGbmSingle(100)
}
