% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{Sharp}
\alias{Sharp}
\title{Calculate Sharp Ratio with stock prices}
\usage{
Sharp(x, rf = 0, n = 10)
}
\arguments{
\item{x}{vector of price}

\item{rf}{risk free rate}

\item{n}{the length of running window}
}
\value{
The sharp ratio series with length the same as x
}
\description{
Calculate sharp ratio of stock with running window.
}
\examples{
date <- as.Date("2015-01-01") + days(0:29)
trade <- rTrade(date)
x <- trade$Close
Sharp(x)
}
