% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos.R
\name{StNormChf}
\alias{StNormChf}
\title{The Characteristic Function of Standard Normal Distribution}
\usage{
StNormChf(u)
}
\arguments{
\item{u}{observation}
}
\value{
The value of Characteristic Function
}
\description{
The Characteristic Function of Standard Normal Distribution
}
\examples{
StNormChf(1)
}
