% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos.R
\name{V_k}
\alias{V_k}
\title{Calculate the V_k Series for Option Pricing with the COS Method}
\usage{
V_k(ValueOption, N, a, b, method = "integrate")
}
\arguments{
\item{ValueOption}{the value function of the option}

\item{N}{the number of cos term for summation}

\item{a}{the lower limit of the truncation interval}

\item{b}{the upper limit of the truncation interval}

\item{method}{how to calculate the integral, one of "integrate" and "jiahe"}
}
\value{
The V_k series
}
\description{
Calculate the V_k Series for Option Pricing with the COS Method
}
\examples{
r <- 0.1
sigmaS0 <- 0.2
tau <- 10
S0 <- 1
K <- 1
mu <- log(S0) + (r - 0.5 * sigmaS0^2) * tau
sigma <- sigmaS0 * sqrt(tau)
a <- -10
b <- 10
N <- 64
ValueOption <- function(x){EuroCallOption(x,K)}
V_k(ValueOption, N, a, b)
}
